Refactor command handling: Remove TP-Link command functions and PineAP detection; add misc CLI handler. Update help commands and improve code organization.
