+++ include/core/cli-handlers/misc-cli-handler.h
@@ -1,0 +1,8 @@
+#pragma once
+
+void handle_help(int argc, char **argv);
+void handle_startwd(int argc, char **argv);
+void handle_scan_ports(int argc, char **argv);
+void handle_web_auth_cmd(int argc, char **argv);
+void handle_pineap_detection(int argc, char **argv);
+void handle_tp_link_test(int argc, char **argv);
